+++ hive_flutter/lib/src/hive_extensions.dart
@@ -5,6 +5,8 @@ extension HiveX on HiveInterface {
   /// Initializes Hive with the path from [getApplicationDocumentsDirectory].
   ///
   /// You can provide a [subDir] where the boxes should be stored.
+  ///
+  /// Also registers the flutter type adapters.
   Future<void> initFlutter([
     String? subDir,
     HiveStorageBackendPreference backendPreference =
@@ -22,5 +24,8 @@ extension HiveX on HiveInterface {
       path,
       backendPreference: backendPreference,
     );
+
+    registerAdapter(ColorAdapter());
+    registerAdapter(TimeOfDayAdapter());
   }
 }