+++ hive/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.4.0
+
+- Adds support for Sets
+- Adds a built in `DurationAdapter`
+- Adds a warning if attempting to register multiple adapters for the same type
+
 ## 2.3.0
 
 - The first release of Hive Community Edition